+++ js/constants.js
@@ -1,0 +1,47 @@
+const PHOTOS = 25;
+
+const MIN_LIKES = 15;
+const MAX_LIKES = 200;
+
+const MIN_COMMENTS = 0;
+const MAX_COMMENTS = 30;
+
+const COMMENTS = 5000;
+
+const MIN_AVATAR = 1;
+const MAX_AVATAR = 6;
+
+const DESCRIPTIONS = [
+  'утро',
+  'созвон',
+  'котик',
+  'пёсик',
+  'снег'
+];
+
+const MESSAGES = [
+  'В целом всё неплохо. Но не всё.',
+  'В целом всё плохо.'
+];
+
+const NAMES = [
+  'Виктория',
+  'Иван',
+  'Вадим',
+  'Елена',
+  'Кирилл'
+];
+
+export {
+  PHOTOS,
+  MIN_LIKES,
+  MAX_LIKES,
+  MIN_COMMENTS,
+  MAX_COMMENTS,
+  COMMENTS,
+  MIN_AVATAR,
+  MAX_AVATAR,
+  DESCRIPTIONS,
+  MESSAGES,
+  NAMES
+};
